Update postinstall.sh
propose to update dir to /Volumes/bigmac, since if instructions have been followed to this point that is where the scripts should be looking. If a user does not switch to this directory then he is borked later in the script with the usage of pwd.

I haven't tested this code to work.